+++ 1.20.4/Common/src/main/java/fuzs/enchantmentcontrol/impl/world/item/enchantment/EnchantmentDataImpl.java
@@ -6,8 +6,10 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import fuzs.enchantmentcontrol.api.v1.data.EnchantmentData;
 import fuzs.enchantmentcontrol.api.v1.data.EnchantmentDataBuilder;
+import fuzs.enchantmentcontrol.api.v1.data.ExpressionEvaluator;
 import fuzs.enchantmentcontrol.impl.EnchantmentControlMod;
 import fuzs.enchantmentcontrol.impl.config.CommonConfig;
+import fuzs.enchantmentcontrol.impl.util.ExpressionEvaluatorImpl;
 import fuzs.puzzleslib.api.config.v3.json.GsonEnumHelper;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
@@ -15,6 +17,7 @@ import net.minecraft.util.GsonHelper;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.*;
 
@@ -23,6 +26,8 @@ public record EnchantmentDataImpl(EnchantmentCategory enchantmentCategory,
                                   EquipmentSlot[] equipmentSlots,
                                   int minLevel,
                                   int maxLevel,
+                                  @Nullable ExpressionEvaluator minCost,
+                                  @Nullable ExpressionEvaluator maxCost,
                                   List<Enchantment> aliases) implements EnchantmentDataBuilder {
     private static final Set<EquipmentSlot> ARMOR_SLOTS = Sets.immutableEnumSet(EquipmentSlot.FEET,
             EquipmentSlot.LEGS,
@@ -37,6 +42,8 @@ public record EnchantmentDataImpl(EnchantmentCategory enchantmentCategory,
                 this.equipmentSlots,
                 this.minLevel,
                 this.maxLevel,
+                this.minCost,
+                this.maxCost,
                 this.aliases
         );
     }
@@ -48,6 +55,8 @@ public record EnchantmentDataImpl(EnchantmentCategory enchantmentCategory,
                 equipmentSlots,
                 this.minLevel,
                 this.maxLevel,
+                this.minCost,
+                this.maxCost,
                 this.aliases
         );
     }
@@ -59,6 +68,8 @@ public record EnchantmentDataImpl(EnchantmentCategory enchantmentCategory,
                 this.equipmentSlots,
                 minLevel,
                 this.maxLevel,
+                this.minCost,
+                this.maxCost,
                 this.aliases
         );
     }
@@ -70,6 +81,32 @@ public record EnchantmentDataImpl(EnchantmentCategory enchantmentCategory,
                 this.equipmentSlots,
                 this.minLevel,
                 maxLevel,
+                this.minCost,
+                this.maxCost,
+                this.aliases
+        );
+    }
+
+    @Override
+    public EnchantmentDataBuilder withMinCost(String minCost) {
+        return new EnchantmentDataImpl(this.enchantmentCategory,
+                this.rarity,
+                this.equipmentSlots,
+                this.minLevel,
+                this.maxLevel, ExpressionEvaluatorImpl.create(minCost.replaceAll("\\s", "")),
+                this.maxCost,
+                this.aliases
+        );
+    }
+
+    @Override
+    public EnchantmentDataBuilder withMaxCost(String maxCost) {
+        return new EnchantmentDataImpl(this.enchantmentCategory,
+                this.rarity,
+                this.equipmentSlots,
+                this.minLevel,
+                this.maxLevel,
+                this.minCost, ExpressionEvaluatorImpl.create(maxCost.replaceAll("\\s", "")),
                 this.aliases
         );
     }
@@ -81,6 +118,8 @@ public record EnchantmentDataImpl(EnchantmentCategory enchantmentCategory,
                 this.equipmentSlots,
                 this.minLevel,
                 this.maxLevel,
+                this.minCost,
+                this.maxCost,
                 List.copyOf(Arrays.asList(aliases))
         );
     }
@@ -101,6 +140,12 @@ public record EnchantmentDataImpl(EnchantmentCategory enchantmentCategory,
         jsonObject.add("slots", equipmentSlots);
         jsonObject.addProperty("min_level", this.minLevel);
         jsonObject.addProperty("max_level", this.maxLevel);
+        if (this.minCost != null) {
+            jsonObject.addProperty("min_cost", this.minCost.getString());
+        }
+        if (this.maxCost != null) {
+            jsonObject.addProperty("max_cost", this.maxCost.getString());
+        }
         JsonArray aliases = new JsonArray();
         for (Enchantment enchantment : this.aliases) {
             aliases.add(BuiltInRegistries.ENCHANTMENT.getKey(enchantment).toString());
@@ -120,6 +165,18 @@ public record EnchantmentDataImpl(EnchantmentCategory enchantmentCategory,
         }
         int minLevel = GsonHelper.getAsInt(jsonObject, "min_level");
         int maxLevel = GsonHelper.getAsInt(jsonObject, "max_level");
+        ExpressionEvaluator minCost;
+        if (jsonObject.has("min_cost")) {
+            minCost = ExpressionEvaluatorImpl.create(GsonHelper.getAsString(jsonObject, "min_cost"));
+        } else {
+            minCost = null;
+        }
+        ExpressionEvaluator maxCost;
+        if (jsonObject.has("max_cost")) {
+            maxCost = ExpressionEvaluatorImpl.create(GsonHelper.getAsString(jsonObject, "max_cost"));
+        } else {
+            maxCost = null;
+        }
         JsonArray aliasesArray = GsonHelper.getAsJsonArray(jsonObject, "aliases", new JsonArray());
         List<Enchantment> aliases = new ArrayList<>();
         for (int i = 0; i < aliasesArray.size(); i++) {
@@ -135,6 +192,8 @@ public record EnchantmentDataImpl(EnchantmentCategory enchantmentCategory,
                 equipmentSlots,
                 minLevel,
                 maxLevel,
+                minCost,
+                maxCost,
                 List.copyOf(aliases)
         );
     }
@@ -152,13 +211,16 @@ public record EnchantmentDataImpl(EnchantmentCategory enchantmentCategory,
                 getAndExpandSlots(enchantment),
                 enchantment.getMinLevel(),
                 enchantment.getMaxLevel(),
+                null,
+                null,
                 Collections.emptyList()
         );
     }
 
     private static EquipmentSlot[] getAndExpandSlots(Enchantment enchantment) {
         EquipmentSlot[] slots = enchantment.slots;
-        if (EnchantmentControlMod.CONFIG.get(CommonConfig.class).unlockArmorEquipmentSlots) {
+        if (!EnchantmentControlMod.CONFIG.getHolder(CommonConfig.class).isAvailable() ||
+                EnchantmentControlMod.CONFIG.get(CommonConfig.class).unlockArmorEquipmentSlots) {
             int armorSlots = 0;
             for (EquipmentSlot slot : slots) {
                 if (slot.isArmor()) armorSlots++;